moved modeshape-jcr-rest-boot project back to modeshape folder from camunda folder.

+++ wcm-bpm-service/spring-commons/spring-boot-commons-non-reactive/spring-boot-commons-service/src/main/java/com/bpwizard/spring/boot/commons/service/security/HttpCookieOAuth2AuthorizationRequestRepository.java
@@ -26,7 +26,7 @@ public class HttpCookieOAuth2AuthorizationRequestRepository implements Authoriza
 	
 	public HttpCookieOAuth2AuthorizationRequestRepository(SpringProperties properties) {
 
-		cookieExpirySecs = properties.getJwt().getShortLivedMillis() / 1000;
+		cookieExpirySecs = (int)properties.getJwt().getShortLivedMillis() / 1000;
 	}
 
 	/**
